install deps
