+++ apps/mobile/package.json
@@ -29,6 +29,7 @@
     "expo-updates": "~0.16.3",
     "lodash.merge": "^4.6.2",
     "lru-cache": "^9.1.1",
+    "mixpanel-react-native": "^2.2.5",
     "nativewind": "^2.0.11",
     "node-html-parser": "^6.1.5",
     "react": "18.2.0",
